pyproject.toml
[project]
name = "syntex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.2",
    "ftfy>=6.3.1",
    "hydra-core>=1.3.2",
    "lightning>=2.5.3",
    "omegaconf>=2.3.0",
    "opencv-python>=4.10.0.84",
    "pillow>=11.1.0",
    "tensorboard>=2.20.0",
    "torch==2.7.0",
    "torchvision>=0.20.1",
    "transformers==4.53.0",
    "wand>=0.6.13",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.4.5",
    "torchinfo>=1.8.0",
]

[tool.uv.pip]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
